+++ barber_v1/assets/font awsome/js/all.js
