+++ barber_v1/index.html
@@ -1,0 +1,59 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <!-- Required meta tags -->
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <link href="https://fonts.googleapis.com/css2?family=Gothic+A1:wght@500&family=Roboto&display=swap" rel="stylesheet">
+<link rel="stylesheet" href="css/style.css">
+    <!-- Bootstrap CSS -->
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css" integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
+
+    <title>Barber landing page</title>
+  </head>
+  <body id="home">
+
+    <section style="background: #162838; " >
+      <div class="thirty">
+        <nav class="navbar navbar-expand-lg sticky-top navbar-light sticky-top">
+          <a class="navbar-brand" href="#"><img src="./assets/images/homepage/Logos.png" alt="" class="img10"></a>
+          <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarTogglerDemo02" aria-controls="navbarTogglerDemo02" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+          </button>
+        
+        
+          <div class="collapse navbar-collapse" id="navbarTogglerDemo02" >
+            <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
+              <li class="nav-item active">
+                <a class="nav-link new9 " href="#home" >Home </a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link new9" href="#about">About</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link new9" href="#services">Services</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link new9" href="#price">Prices</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link new9" href="#contact">Contact</a>
+              </li>
+            </ul>
+            
+          </div>
+        </nav>
+      </div>
+      
+    </section>
+    
+</main>
+  
+      
+    <!-- Optional JavaScript -->
+    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
+    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js" integrity="sha384-OgVRvuATP1z7JjHLkuOU7Xw704+h835Lr+6QL9UvYjZE3Ipu6Tp75j7Bh/kR0JKI" crossorigin="anonymous"></script>
+  </body>
+</html>
